clean up code, delete service worker, and fix bug in date input --WORKING STATE--
The application is currently in a working state. When the user makes an input, that input will be 'set' in chrome's local storage API. A message will be sent to the background script to initiate authentication flow. The authentication flow will open a popup (for Meetup.com) and, once authenticated, a redirect uri will be given to the client and exchanged for an access token. The access token can then be used to make API calls.