add reset relay
The content script fires a message to reset the text field after the application sends a message called 'popUpInit' in the componentDIdMount lifecycle method. This will work just fine for the application registered in popup.html, but the application that has been injected into the content  of the page will not here this. In order to correct this issue a relay message needs to be configured in the background script that will listen for the resetTextField message and then send the correct corresponding message back to the correct tab in the content script.