fix bug where the data object coming up as undefined
when the data object turns up undefined, then the promise will continue it's fulfillments with an undefined value and break along the way. I could forcefully break the promise and return an error but this would be a hacky solution and a bit more verbose than neccessary. Instead I went ahead and added a dummy object and conditional logic that would assign the parseddata obj to the dummy object just in case nothing was found